requirements.txt
absl-py==1.0.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
gcloud==0.18.3
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
oauth2client==4.1.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyparsing==3.0.8
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
